Refactor: Extract internal modules for better organization
- Moved Git operations, stats, and display functions to
- src/internal/git, src/internal/stats, and src/internal/display.
- Added src/internal/utils for utilities. Updated README.md with
- CodeRabbit badge to enhance documentation.
